requirements.txt
annotated-types==0.7.0
anyio==4.7.0
astroid==3.3.6
autopep8==2.3.1
certifi==2024.12.14
click==8.1.7
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==5.13.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mysql-connector-python==9.1.0
numpy==2.2.0
platformdirs==4.3.6
pyarrow==18.1.0
pycodestyle==2.12.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pylance==0.20.0
pylint==3.3.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
tomlkit==0.13.2
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
